chore(cleanup): improved the module folder structure, finalized initial test envrionment, toggled welcome on load
