+++ module/api/dice/meta-rolld10.mjs
@@ -1,4 +1,4 @@
-import { metaSheetRefresh } from "../helpers/metahelpers.mjs";
+import { metaSheetRefresh } from "../../helpers/metahelpers.mjs";
 /**
  * metaRolld10 handles the rolling of d10 dice for a given actor and purpose.
  *
@@ -163,172 +163,5 @@ export async function metaRolld10ReRoll(event) {
 	} else {
 		ui.notifications.warn(actor.name + " does not have enough Destiny to spend for reroll!");
 		metanthropes.utils.metaLog(1, "metaRolld10ReRoll", "Not enough Destiny to spend", "OR", "destinyReRoll is not allowed");
-	}
-}
-
-/**
- * metaHungerRoll handles the rolling a simple d100 to check if the actor can beat their Hunger effect.
- */
-export async function metaHungerRoll(actor, hungerLevel) {
-	let hungerTarget = 0;
-	let hungerEffect = 0; //! not needed???
-	let hungerMessage = null;
-	if (hungerLevel === 1) {
-		hungerEffect = 10;
-		hungerTarget = 90;
-	} else if (hungerLevel === 2) {
-		hungerEffect = 25;
-		hungerTarget = 75;
-	} else if (hungerLevel === 3) {
-		hungerEffect = 50;
-		hungerTarget = 50;
-	} else if (hungerLevel === 4) {
-		hungerEffect = 75;
-		hungerTarget = 25;
-	} else if (hungerLevel === 5) {
-		hungerEffect = 90;
-		hungerTarget = 10;
-	} else {
-		metanthropes.utils.metaLog(5, "metaHungerRoll", "Hunger Level is not valid:", hungerLevel);
-		return;
-	}
-	const hungerRoll = await new Roll("1d100").evaluate();
-	const hungerRollResult = hungerRoll.total;
-	hungerMessage = `Rolls to beat Hunger 💀 Condition Level ${hungerLevel} and gets a result of ${hungerRollResult} (needed ${hungerTarget} or less).<br><br>`;
-	if (hungerRollResult > hungerTarget) {
-		hungerMessage += `It is a 🟥 Failure!<br><br>${actor.name} is too hungry and can't act!`;
-		//? Button to re-roll Hunger using destiny
-		const currentDestiny = Number(actor.system.Vital.Destiny.value);
-		hungerMessage += `<hr />${actor.name} has ${currentDestiny} 🤞 Destiny remaining.<br>`;
-		if (currentDestiny > 0) {
-			hungerMessage += `<div class="hide-button hidden"><br><button class="metanthropes-main-chat-button hunger-reroll" 
-			data-actoruuid="${actor.uuid}" data-hunger-level="${hungerLevel}"
-			>Spend 🤞 Destiny to reroll</button><br></div><br>`;
-		}
-	} else {
-		hungerMessage += `It is a 🟩 Success!<hr />${actor.name} has overcome Hunger!<br><br>`;
-		await actor.setFlag("metanthropes", "hungerRollResult", true);
-		const MetaRollBeforeHungerCheck = await actor.getFlag("metanthropes", "MetaRollBeforeHungerCheck");
-		metanthropes.utils.metaLog(
-			3,
-			"metaHungerRoll",
-			"Engaging metaRoll with:",
-			actor,
-			MetaRollBeforeHungerCheck.action,
-			MetaRollBeforeHungerCheck.stat,
-			MetaRollBeforeHungerCheck.isCustomRoll,
-			MetaRollBeforeHungerCheck.destinyCost,
-			MetaRollBeforeHungerCheck.itemName
-		);
-		metanthropes.dice.metaRoll(
-			actor,
-			MetaRollBeforeHungerCheck.action,
-			MetaRollBeforeHungerCheck.stat,
-			MetaRollBeforeHungerCheck.isCustomRoll,
-			MetaRollBeforeHungerCheck.destinyCost,
-			MetaRollBeforeHungerCheck.itemName
-		);
-		await actor.unsetFlag("metanthropes", "MetaRollBeforeHungerCheck");
-	}
-	hungerRoll.toMessage({
-		speaker: ChatMessage.getSpeaker({ actor: actor }),
-		flavor: hungerMessage,
-		rollMode: game.settings.get("core", "rollMode"),
-		flags: { "metanthropes": { actoruuid: actor.uuid } },
-	});
-}
-
-/**
- * Hunger ReRoll
- */
-export async function metaHungerReRoll(event) {
-	event.preventDefault();
-	const button = event.target;
-	const actoruuid = button.dataset.actoruuid;
-	const hungerLevel = parseInt(button.dataset.hungerLevel);
-	const actor = await fromUuid(actoruuid);
-	let currentDestiny = Number(actor.system.Vital.Destiny.value);
-	if (currentDestiny > 0) {
-		currentDestiny--;
-		await actor.update({ "system.Vital.Destiny.value": Number(currentDestiny) });
-		metaHungerRoll(actor, hungerLevel);
-	} else {
-		ui.notifications.warn(actor.name + " does not have enough Destiny to spend for reroll!");
-		metanthropes.utils.metaLog(3, "metaHungerReRoll", "Not enough Destiny to spend");
-	}
-}
-
-/**
- * metaCoverRoll handles the rolling a simple d100 to check if the actor can find Cover.
- * 
- * @param {*} actor 
- * @param {*} coverType 
- * @param {*} coverValue 
- * @returns 
- */
-export async function metaCoverRoll(actor, coverType, coverValue) {
-	let coverMessage = null;
-	let coverTarget = null;
-	if (coverValue === 0) {
-		ui.notifications.warn(actor.name + " does not have any " + coverType + " Cover to roll for!");
-		return;
-	} else if (coverValue === 10) {
-		coverTarget = 10;
-	} else if (coverValue === 25) {
-		coverTarget = 25;
-	} else if (coverValue === 50) {
-		coverTarget = 50;
-	} else if (coverValue === 75) {
-		coverTarget = 75;
-	} else if (coverValue === 90) {
-		coverTarget = 90;
-	} else {
-		metanthropes.utils.metaLog(5, "metaCoverRoll", "Cover Value is not valid:", coverValue);
-		return;
-	}
-	const coverRoll = await new Roll("1d100").evaluate();
-	const coverRollResult = coverRoll.total;
-	coverMessage = `Rolls to find ${coverType} Cover, with ${coverValue}% and gets a result of ${coverRollResult} (needed ${coverTarget} or less).<br><br>`;
-	if (coverRollResult > coverTarget) {
-		coverMessage += `It is a 🟥 Failure!<br><br>${actor.name} can't find Cover!`;
-		//? Button to re-roll Cover using destiny
-		const currentDestiny = Number(actor.system.Vital.Destiny.value);
-		coverMessage += `<hr />${actor.name} has ${currentDestiny} 🤞 Destiny remaining.<br>`;
-		if (currentDestiny > 0) {
-			coverMessage += `<div class="hide-button hidden"><br><button class="metanthropes-main-chat-button cover-reroll" 
-			data-actoruuid="${actor.uuid}" data-cover-value="${coverValue}" data-type="${coverType}"
-			>Spend 🤞 Destiny to reroll</button><br></div><br>`;
-		}
-	} else {
-		coverMessage += `It is a 🟩 Success!<hr />${actor.name} found ${coverType} Cover!<br><br>`;
-	}
-	coverRoll.toMessage({
-		speaker: ChatMessage.getSpeaker({ actor: actor }),
-		flavor: coverMessage,
-		rollMode: game.settings.get("core", "rollMode"),
-		flags: { "metanthropes": { actoruuid: actor.uuid } },
-	});
-}
-
-/**
- * Cover ReRoll
- * 
- * @param {*} event 
- */
-export async function metaCoverReRoll(event) {
-	event.preventDefault();
-	const button = event.target;
-	const actoruuid = button.dataset.actoruuid;
-	const actor = await fromUuid(actoruuid);
-	const coverType = button.dataset.type;
-	const coverValue = parseInt(button.dataset.coverValue);
-	let currentDestiny = Number(actor.system.Vital.Destiny.value);
-	if (currentDestiny > 0) {
-		currentDestiny--;
-		await actor.update({ "system.Vital.Destiny.value": Number(currentDestiny) });
-		metaCoverRoll(actor, coverType, coverValue);
-	} else {
-		ui.notifications.warn(actor.name + " does not have enough Destiny to spend for reroll!");
-		metanthropes.utils.metaLog(3, "metaCoverReRoll", "Not enough Destiny to spend");
 	}
 }
